Rename python files (#19)
