+++ tests/test_CG_05_GET_buscar_customer_group.py
@@ -1,7 +1,73 @@
 import pytest
 import requests
+import jsonschema
 
 from src.singleton import Singleton
+
+
+@pytest.mark.functional
+@pytest.mark.regression
+def test_CG05TC4_GET_verificar_que_retorna_un_error_al_buscar_los_10_primeros_customer_group_con_el_metodo_POST(get_token_login):
+    token = get_token_login
+    url = f"{Singleton.get_base_url()}/rest/default/V1/customerGroups/search"
+
+    headers = {
+        'Authorization': f'Bearer {token}',
+        'Content-Type': 'application/json',
+    }
+
+    payload = {}
+
+    response = requests.post(url, headers=headers, json=payload)
+    response_data = response.json()
+
+    assert response.status_code == 404, "NO FOUND 404"
+
+
+@pytest.mark.functional
+@pytest.mark.regression
+def test_CG05TC5_GET_verificar_que_retorna_un_error_al_buscar_los_10_primeros_customer_group_con_el_metodo_PUT(get_token_login):
+    token = get_token_login
+    url = f"{Singleton.get_base_url()}/rest/default/V1/customerGroups/search"
+
+    headers = {
+        'Authorization': f'Bearer {token}',
+        'Content-Type': 'application/json',
+    }
+
+    payload = {}
+
+    response = requests.put(url, headers=headers, json=payload)
+    response_data = response.json()
+
+    assert response.status_code == 400, "NO FOUND 400"
+
+
+@pytest.mark.functional
+@pytest.mark.regression
+def test_CG05TC6_GET_verificar_que_retorna_un_error_al_buscar_los_10_primeros_customer_group_con_el_metodo_DELETE(get_token_login):
+    token = get_token_login
+    url = f"{Singleton.get_base_url()}/rest/default/V1/customerGroups/search"
+
+    headers = {
+        'Authorization': f'Bearer {token}',
+        'Content-Type': 'application/json',
+    }
+
+    payload = {}
+
+    response = requests.delete(url, headers=headers, json=payload)
+    response_data = response.json()
+
+    assert response.status_code == 400, "NO FOUND 400"
+
+def test_schema_verificar_obtencion_exitosa_de_los_customer_groups(get_body_obtain_first_10_customer_groups):
+    response_data = get_body_obtain_first_10_customer_groups
+    schema = Singleton.read_schema_json_file('get_obtain_first_10_customer_groups.json')
+    try:
+        jsonschema.validate(instance=response_data, schema=schema)
+    except jsonschema.exceptions.ValidationError as err:
+        pytest.fail(f'JSON schema dont match {err}')
 
 
 @pytest.mark.smoke
