+++ tests/test_CG_06_GET_obtener_customer_group_por_id.py
@@ -1,7 +1,7 @@
 import pytest
 import requests
 
-from src.assertions.asserions_schema import assert_schemas
+from src.assertions.assertions_schema import assert_schemas
 from src.singleton import Singleton
