+++ tests/test_C_2_POST_crear_nuevo_customer.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.assertions.asserions_schema import assert_schemas
+from src.assertions.assertions_schema import assert_schemas
 
 
 @pytest.mark.smoke
